Revert "Merge pull request #107 from alexanderwanyoike/feat/perform-card-19-live-notes-automation"
This reverts commit bc0865d23126bacc8e1a21d118509566479fe8df, reversing
changes made to f2748b06278e1aef86950f236046b3c8775c0dcb.
