+++ crates/vibez-engine/src/engine_mute_automation_tests.rs
@@ -2,10 +2,9 @@ use std::sync::Arc;
 
 use vibez_core::audio_buffer::DecodedAudio;
 use vibez_core::automation::{AutomationLane, AutomationPoint, AutomationTarget};
-use vibez_core::id::{ClipId, SectionId, TrackId};
+use vibez_core::id::{ClipId, TrackId};
 
 use super::*;
-use crate::playback_source::{EngineClip, PreparedPlaybackSource, PreparedSectionPlaybackSource};
 
 fn constant_audio(frames: usize) -> Arc<DecodedAudio> {
     Arc::new(DecodedAudio {
@@ -107,139 +106,4 @@ fn manual_mute_override_holds_until_automation_is_reenabled() {
         0.0,
         "re-enable must return control to the lane"
     );
-}
-
-#[test]
-fn live_pan_gesture_overrides_arrange_baseline_then_yields_back() {
-    let (mut engine, mut commands, mut events) = AudioEngine::new();
-    let track_id = TrackId::new();
-    add_constant_track(&mut commands, track_id);
-    let mut lane = AutomationLane::new(AutomationTarget::TrackPan);
-    lane.insert_point(AutomationPoint {
-        beat: 0.0,
-        value: 0.0,
-        curve: 0.0,
-    });
-    commands
-        .push(EngineCommand::SetAutomationLane { track_id, lane })
-        .unwrap();
-    commands.push(EngineCommand::Play).unwrap();
-    let mut output = vec![0.0; 16 * 2];
-    engine.process(&mut output, 2);
-    assert!(output[0].abs() > output[1].abs());
-
-    commands
-        .push(EngineCommand::UpdateAutomationGesture {
-            track_id,
-            target: AutomationTarget::TrackPan,
-            normalized_value: 1.0,
-            begin: true,
-        })
-        .unwrap();
-    commands
-        .push(EngineCommand::SetTrackPan(track_id, 1.0))
-        .unwrap();
-    output.fill(0.0);
-    engine.process(&mut output, 2);
-    assert!(output[1].abs() > output[0].abs());
-
-    commands
-        .push(EngineCommand::EndAutomationGesture {
-            track_id,
-            target: AutomationTarget::TrackPan,
-        })
-        .unwrap();
-    output.fill(0.0);
-    engine.process(&mut output, 2);
-    assert!(output[0].abs() > output[1].abs());
-
-    let gestures: Vec<_> = std::iter::from_fn(|| events.pop().ok())
-        .filter_map(|event| match event {
-            EngineEvent::AutomationGestureChanged {
-                phase,
-                normalized_value,
-                ..
-            } => Some((phase, normalized_value)),
-            _ => None,
-        })
-        .collect();
-    assert_eq!(
-        gestures,
-        [
-            (crate::events::AutomationGesturePhase::Begin, 1.0),
-            (crate::events::AutomationGesturePhase::End, 0.0),
-        ]
-    );
-}
-
-#[test]
-fn perform_gesture_uses_the_active_section_lane_not_arrange_automation() {
-    let (mut engine, mut commands, _events) = AudioEngine::new();
-    let track_id = TrackId::new();
-    commands.push(EngineCommand::SetSampleRate(100)).unwrap();
-    commands.push(EngineCommand::SetBpm(60.0)).unwrap();
-    commands
-        .push(EngineCommand::AddTrack(track_id, "Section pan".into()))
-        .unwrap();
-    let mut section_pan = AutomationLane::new(AutomationTarget::TrackPan);
-    section_pan.insert_point(AutomationPoint {
-        beat: 0.0,
-        value: 0.0,
-        curve: 0.0,
-    });
-    commands
-        .push(EngineCommand::LaunchSection(Box::new(
-            PreparedSectionPlaybackSource::new(
-                SectionId::new(),
-                4.0,
-                true,
-                vec![(
-                    track_id,
-                    PreparedPlaybackSource::new(
-                        vec![EngineClip {
-                            id: ClipId::new(),
-                            audio: constant_audio(512),
-                            position: 0,
-                            source_offset: 0,
-                            duration: 512,
-                            loop_enabled: false,
-                            loop_start: 0,
-                            loop_end: 0,
-                        }],
-                        Vec::new(),
-                        vec![section_pan],
-                    ),
-                )],
-            ),
-        )))
-        .unwrap();
-
-    let mut output = vec![0.0; 16 * 2];
-    engine.process(&mut output, 2);
-    assert!(output[0].abs() > output[1].abs());
-
-    commands
-        .push(EngineCommand::UpdateAutomationGesture {
-            track_id,
-            target: AutomationTarget::TrackPan,
-            normalized_value: 1.0,
-            begin: true,
-        })
-        .unwrap();
-    commands
-        .push(EngineCommand::SetTrackPan(track_id, 1.0))
-        .unwrap();
-    output.fill(0.0);
-    engine.process(&mut output, 2);
-    assert!(output[1].abs() > output[0].abs());
-
-    commands
-        .push(EngineCommand::EndAutomationGesture {
-            track_id,
-            target: AutomationTarget::TrackPan,
-        })
-        .unwrap();
-    output.fill(0.0);
-    engine.process(&mut output, 2);
-    assert!(output[0].abs() > output[1].abs());
 }
